Move non-ticket views to orders/views.py
